spk/Resources/NemotronRuntime/requirements.txt
numpy<2
packaging
setuptools<81
wheel
Cython<3.1
nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git@main
